final commit!

+++ The_Book_Store/Admin/FormStockIn.cs
@@ -132,16 +132,18 @@ namespace The_Book_Store.Admin
 
         private void BtnCancel_Click(object sender, EventArgs e)
         {
-            if (MessageBox.Show("Are you sure you want to clear these records?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            if (dataGridViewStockIn.Rows.Count > 0)
             {
-                dataGridViewStockIn.Rows.Clear();
-                cn.Open();
-                cm = new SqlCommand("DELETE FROM tblStockIn", cn);
-                cm.ExecuteNonQuery();
-                cn.Close();
-                CheckForChangesAndUpdateFlag();
+                if (MessageBox.Show("Are you sure you want to clear these records?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                {
+                    dataGridViewStockIn.Rows.Clear();
+                    cn.Open();
+                    cm = new SqlCommand("DELETE FROM tblStockIn", cn);
+                    cm.ExecuteNonQuery();
+                    cn.Close();
+                    CheckForChangesAndUpdateFlag();
+                }
             }
-
         }
 
         private void FormStockIn_Load(object sender, EventArgs e)
@@ -158,6 +160,7 @@ namespace The_Book_Store.Admin
                 {
                     e.Cancel = true;
                     SaveData();
+                    this.Close();
                 } 
                 else if(result == DialogResult.No)
                 {
